backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-SocketIO
geopy
eventlet
psycopg2-binary
python-dotenv
pytest
pytest-cov
